add delele opt_cache for trt_split_converter_test
