+++ paddle/fluid/inference/tests/api/trt_dynamic_shape_ernie_serialize_deserialize_test.h
@@ -12,7 +12,6 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License. */
 #pragma once
-#include <dirent.h>
 #include <glog/logging.h>
 #include <gtest/gtest.h>
 #include <unistd.h>
@@ -26,22 +25,6 @@ limitations under the License. */
 
 namespace paddle {
 namespace inference {
-
-static int DeleteCache(std::string path) {
-  DIR* dir = opendir(path.c_str());
-  if (dir == NULL) return 0;
-  struct dirent* ptr;
-  while ((ptr = readdir(dir)) != NULL) {
-    if (std::strcmp(ptr->d_name, ".") == 0 ||
-        std::strcmp(ptr->d_name, "..") == 0) {
-      continue;
-    } else if (ptr->d_type == 8) {
-      std::string file_rm = path + "/" + ptr->d_name;
-      return remove(file_rm.c_str());
-    }
-  }
-  return 0;
-}
 
 static void run(const AnalysisConfig& config, std::vector<float>* out_data) {
   auto predictor = CreatePaddlePredictor(config);
@@ -111,7 +94,7 @@ static void trt_ernie(bool with_fp16, std::vector<float> result) {
   // Delete serialization cache to perform serialization first rather than
   // deserialization.
   std::string opt_cache_dir = FLAGS_infer_model + "/_opt_cache";
-  DeleteCache(opt_cache_dir);
+  delete_cache_files(opt_cache_dir);
 
   SetConfig(&config, model_dir, true /* use_gpu */);
